+++ Source/FunicularSwitch.Generators/ExtendMonad/ExtendMonadAttribute.cs
@@ -1,6 +1,6 @@
 using DarkLink.RoslynHelpers;
 
-namespace FunicularSwitch.Generators.Monad;
+namespace FunicularSwitch.Generators.ExtendMonad;
 
 [GenerateAttribute(AttributeTargets.Class | AttributeTargets.Struct, Namespace = "FunicularSwitch.Generators", Inherited = false)]
 public partial record ExtendMonadAttribute(bool __ignore_me = false);
