+++ Source/FunicularSwitch.Generators/ExtendMonad/Generator.cs
@@ -1,7 +1,7 @@
 using FunicularSwitch.Generators.Generation;
 using Microsoft.CodeAnalysis;
 
-namespace FunicularSwitch.Generators.Monad;
+namespace FunicularSwitch.Generators.ExtendMonad;
 
 internal class Generator
 {
